api/requirements.txt
redis==5.0.3
elasticsearch[async]==7.9.1
fastapi==0.110.1
orjson==3.10.0
pydantic==2.6.4
pydantic-settings~=2.2.1
python-dotenv~=1.0.1
uvicorn==0.29.0
uvloop==0.19.0 ; sys_platform != "win32" and implementation_name == "cpython"
gunicorn==22.0.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0 
backoff~=2.2.1
aiohttp==3.9.5
sentry-sdk[fastapi]
motor~=3.5.1
jwcrypto~=1.5.6